fix: remove unwanted social buttons
Several places misuses the ChannelSocials for the block and fav functionality in it. IMO they should be split. Anyway, for now I just correctly hide them where they should not be.

For the hover effect on the channel chips visible in collabs, we show only fav.

For the report button we show only block.

It looks like WatchMentions is currently unused, but I updated it as well.